refacotor tests
