+++ state_beacon/test/src/async_value_test.dart
