+++ state_beacon/test/src/batch_test.dart
