+++ state_beacon/test/src/beacons/value_notifier_test.dart
