+++ state_beacon/test/src/custom_types_test.dart
