+++ state_beacon/test/src/extension_test.dart
