+++ state_beacon/test/src/flutter_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:state_beacon/state_beacon.dart';
 
-import 'common.dart';
+import '../common.dart';
 
 Future<String> counterFuture(int count) async {
   if (count > 3) {
